many unit test updates and code improvements
worked on floret channel unit tests

further breakout of tests
