+++ lib/channel/floret-channel.js
@@ -17,7 +17,6 @@ let FloretChannel;
   const _channelEndpoint = Symbol('channelEndpoint');
   const _handler = Symbol('handler');
   const _broadcaster = Symbol('broadcaster');
-  const _promise = Symbol('promise');
 
   FloretChannel = class FloretChannelClass {
     constructor({
@@ -33,11 +32,10 @@ let FloretChannel;
       this[_subject] = new Rx.Subject();
       this[_subscribers] = {};
       this[_broadcaster] = broadcaster || this.defaultBroadcaster;
-      this[_promise] = rp;
     }
 
     subscribe(subscriber) {
-      if (!this[_subscribers][subscriber.name] && subscriber.endpoint) {
+      if (!this[_subscribers][subscriber.name]) {
         const name = subscriber.name;
         const endpoint = subscriber.endpoint;
 
@@ -48,15 +46,15 @@ let FloretChannel;
 
         this[_subscribers][subscriber.name] = subscriber;
         return subscriber;
+      } else {
+        return this[_subscribers][subscriber.name];
       }
     }
 
     unsubscribe(subscriberName) {
-      if (this[_subscribers][subscriberName]) {
-        this[_subscribers][subscriberName].observer.unsubscribe();
-        this[_subscribers][subscriberName] = {};
-        delete this[_subscribers][subscriberName];
-      }
+      this[_subscribers][subscriberName].observer.unsubscribe();
+      this[_subscribers][subscriberName] = {};
+      delete this[_subscribers][subscriberName];
     }
 
     broadcast(payload, sender = this.name, trackingId) {
@@ -137,15 +135,17 @@ let FloretChannel;
 
     get defaultBroadcaster() {
       return {
-        emit: this[_promise] ? this[_promise].post : rp.post,
+        emit: rp.post,
       };
     }
 
     get handler() {
+      console.log('getting handler')
       return this[_handler];
     }
 
     set handler(hdlr) {
+      console.log('setting handler')
       this[_handler] = hdlr;
     }
 
@@ -158,26 +158,23 @@ let FloretChannel;
         sender, name, channel, payload, trackingId,
       });
 
-      if (endpoint) {
-        const options = {
-          header: {
-            'Content-Type': 'application/json',
-          },
-          method: 'POST',
-          body: pkg.toJSON(),
-          uri: endpoint,
-          json: true,
-        };
+      const options = {
+        header: {
+          'Content-Type': 'application/json',
+        },
+        method: 'POST',
+        body: pkg.toJSON(),
+        uri: endpoint,
+        json: true,
+      };
 
-        // purposefuly do not wait
-        broadcaster.emit(options).then(() => {
-          console.log(`broadcast complete: ${endpoint}`);
-        }).catch((e) => {
-          console.log(`caught broadcast error ${e.message}`);
-          console.log(endpoint);
-          console.log(JSON.stringify(pkg.toJSON()));
-        });
-      }
+      // purposefuly do not wait
+      return broadcaster.emit(options).then(() => {
+        console.log(`broadcast complete: ${endpoint}`);
+      }).catch((e) => {
+          // suppress
+          return;
+      });
     }
 
     config() {
