+++ lib/floret.js
@@ -623,10 +623,11 @@ let FloretClass;
         uri: `${this.gateway.proxyURL}/api-doc/specs`,
         method: 'GET',
       };
+      console.log('about to pusblish')
 
       // notify api-doc service to refresh spec list
       await this.gateway.publishAPISpec(options).then(() => {
-
+        console.log('done publish')
       }).catch((e) => {
         console.log(`Default document service not found (${this.gateway.proxyURL}/api-doc/specs)`);
         console.log(e.message);
@@ -636,7 +637,7 @@ let FloretClass;
         path: '/api-spec.json',
         methods: ['GET', 'POST', 'DELETE'],
       }];
-
+        console.log('done')
       return routes;
     }
 
@@ -648,8 +649,6 @@ let FloretClass;
       const channelConfig = {
         name: 'api-spec',
         description: 'api spec',
-        hostURL: this.host,
-        hostPort: this.port,
         uri: '/api-spec',
         endpoint: `${this.url}${this.uri}/api-spec`,
       };
@@ -938,6 +937,7 @@ let FloretClass;
          */
     async deleteChannel(channel) {
       const channelInfo = channel.config();
+      delete this[_channels][channel.name];
 
       this[_gateway].deleteChannelAPI(channelInfo.apiName).then(() => {
         for (let i = 0; i < this[_channels].length; i++) {
