+++ test/stubs/gateway.stubs.js
@@ -86,22 +86,6 @@ sinon.stub(floret.gateway,'register').resolves(() =>{
 
   sinon.stub(floret.gateway, 'getServiceByName').callsFake((obj) => {
     console.log('getServiceByName stub called');
-    console.log('returning, ', JSON.stringify({
-      host: '192.168.1.68',
-      created_at: 1534218595,
-      connect_timeout: 60000,
-      id: '9c00be92-7ae8-4748-bd06-serviceIdMock',
-      protocol: 'http',
-      name: 'mock-floret',
-      read_timeout: 60000,
-      port: 8082,
-      path: '/',
-      updated_at: 1534218595,
-      retries: 5,
-      write_timeout: 60000,
-
-    }));
-
     return {
       host: '192.168.1.68',
       created_at: 1534218595,
@@ -120,7 +104,7 @@ sinon.stub(floret.gateway,'register').resolves(() =>{
   });
 
   sinon.stub(floret.gateway, 'getServiceRoutes').callsFake((name) => {
-    console.log(`getServiceRoutes stub called: ${name}`);
+    console.log(`getServiceRoutes stub`);
     if (name.indexOf('-channel') > -1) {
       console.log('channel routes');
       return [
@@ -321,57 +305,69 @@ sinon.stub(floret.gateway,'register').resolves(() =>{
     },
   }));
 
-  /*
-sinon.stub(floret.gateway,'apiRequestByName').resolves(() =>{
-    console.log('apiRequestByName stub called')
-});
-
-sinon.stub(floret.gateway,'apiRequestByURI').resolves(() =>{
-    console.log('apiRequestByURI stub called')
-});
-
-sinon.stub(floret.gateway,'getAPIsWithUpstreamURL').resolves(() =>{
-    console.log('getAPIsWithUpstreamURL stub called')
-});
-
-sinon.stub(floret.gateway,'addAPI').resolves(() =>{
-    console.log('addAPI stub called')
-
-});
-
-sinon.stub(floret.gateway,'deleteAPI').resolves(() =>{
+  sinon.stub(floret.gateway,'deleteAPI').resolves(() =>{
     console.log('deleteAPI stub called')
-});
+  });
 
-sinon.stub(floret.gateway,'subscribeTo').resolves( () => {
-    console.log('subscribeTo stub called');
-});
+  sinon.stub(floret.gateway,'createChannelAPI').resolves(() =>{
+    console.log('createChannelAPI stub called')
+  });
 
-sinon.stub(floret.gateway,'unsubscribe').resolves(() =>{
-    console.log('unsubscribe stub called')
-});
-sinon.stub(floret.gateway,'loadSubscribers').resolves(() =>{
-    console.log('loadSubscribers stub called')
-});
+    sinon.stub(floret.gateway,'createChannelService').resolves(() =>{
+        console.log('createChannelService stub called')
+    });
 
-sinon.stub(floret.gateway,'discoverAPISpecs').callsFake(() =>{
-    console.log('apispecs stub')
-    return [{"name": "apiSpec1"}];
-});
-sinon.stub(floret.gateway,'').callsFake(() =>{
-    console.log(' stub called')
-});
-sinon.stub(floret.gateway,'').callsFake(() =>{
-    console.log(' stub called')
-});
-sinon.stub(floret.gateway,'').callsFake(() =>{
-    console.log(' stub called')
-});
-sinon.stub(floret.gateway,'').callsFake(() =>{
-    console.log(' stub called')
-});
-sinon.stub(floret.gateway,'').callsFake(() =>{
-    console.log(' stub called')
-});
-*/
+    /*
+  sinon.stub(floret.gateway,'apiRequestByName').resolves(() =>{
+      console.log('apiRequestByName stub called')
+  });
+
+  sinon.stub(floret.gateway,'apiRequestByURI').resolves(() =>{
+      console.log('apiRequestByURI stub called')
+  });
+
+  sinon.stub(floret.gateway,'getAPIsWithUpstreamURL').resolves(() =>{
+      console.log('getAPIsWithUpstreamURL stub called')
+  });
+
+  sinon.stub(floret.gateway,'addAPI').resolves(() =>{
+      console.log('addAPI stub called')
+
+  });
+
+  sinon.stub(floret.gateway,'deleteAPI').resolves(() =>{
+      console.log('deleteAPI stub called')
+  });
+
+  sinon.stub(floret.gateway,'subscribeTo').resolves( () => {
+      console.log('subscribeTo stub called');
+  });
+
+  sinon.stub(floret.gateway,'unsubscribe').resolves(() =>{
+      console.log('unsubscribe stub called')
+  });
+  sinon.stub(floret.gateway,'loadSubscribers').resolves(() =>{
+      console.log('loadSubscribers stub called')
+  });
+
+  sinon.stub(floret.gateway,'discoverAPISpecs').callsFake(() =>{
+      console.log('apispecs stub')
+      return [{"name": "apiSpec1"}];
+  });
+  sinon.stub(floret.gateway,'').callsFake(() =>{
+      console.log(' stub called')
+  });
+  sinon.stub(floret.gateway,'').callsFake(() =>{
+      console.log(' stub called')
+  });
+  sinon.stub(floret.gateway,'').callsFake(() =>{
+      console.log(' stub called')
+  });
+  sinon.stub(floret.gateway,'').callsFake(() =>{
+      console.log(' stub called')
+  });
+  sinon.stub(floret.gateway,'').callsFake(() =>{
+      console.log(' stub called')
+  });
+  */
 };
